working with User functions

+++ BACK-END/Crowdfunding/src/main/java/com/s8/Crowdfunding/repository/DonationRepository.java
@@ -9,13 +9,15 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DonationRepository extends JpaRepository<Donation, Integer> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
 
     @Query("SELECT COALESCE(SUM(d.amount), 0) FROM Donation d WHERE d.project.projectId = :id")
     Double sumOfAllDonationsById(@Param("id") Long id);
 
-     List<Donation> findDonationsByDonor_UserIdAndStatusEquals(long userId, String status);
-     List<Donation> findDonationsByProjectProjectIdAndStatusEquals(long projectId, String status);
+    List<Donation> findDonationsByDonor_UserIdAndStatusEquals(long userId, String status);
+
+    List<Donation> findDonationsByProjectProjectIdAndStatusEquals(long projectId, String status);
+
     @Query("SELECT COALESCE(SUM(d.amount), 0) FROM Donation d WHERE d.project.projectId = :projectId")
     Double sumOfAllDonationsByProjectId(@Param("projectId") long projectId);
 }
