+++ BACK-END/Crowdfunding/src/main/java/com/s8/Crowdfunding/repository/UserRepository.java
@@ -10,15 +10,16 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Integer> {
-    @Query("SELECT p FROM Project p WHERE p.status = :status AND p.user.userId = :userId")
-    List<Project> findUserByUserIdAndStatus(@Param("status") String status, @Param("userId") Long userId);
+public interface UserRepository extends JpaRepository<Users, Long> {
+        @Query("SELECT p FROM Project p WHERE p.status = :status AND p.user.userId = :userId")
+        List<Project> findUserByUserIdAndStatus(@Param("status") String status, @Param("userId") Long userId);
 
-    @Query("SELECT p FROM Project p WHERE p.status = :status AND p.user.userId = :userId " +
-            "AND (SELECT COALESCE(SUM(d.amount),0) FROM Donation d WHERE d.project = p) < p.goalAmount")
-    List<Project> listOfApprovedAndGoalNotReachedProject(@Param("status") String status, @Param("userId") Long userId);
+        @Query("SELECT p FROM Project p WHERE p.status = :status AND p.user.userId = :userId " +
+                        "AND (SELECT COALESCE(SUM(d.amount),0) FROM Donation d WHERE d.project = p) < p.goalAmount")
+        List<Project> listOfApprovedAndGoalNotReachedProject(@Param("status") String status,
+                        @Param("userId") Long userId);
 
-    @Query("SELECT p FROM Project p WHERE p.status = :status AND p.user.userId = :userId " +
-            "AND (SELECT COALESCE(SUM(d.amount),0) FROM Donation d WHERE d.project = p) >= p.goalAmount")
-    List<Project> listOfApprovedAndGoalReachedProject(@Param("status") String status, @Param("userId") Long userId);
+        @Query("SELECT p FROM Project p WHERE p.status = :status AND p.user.userId = :userId " +
+                        "AND (SELECT COALESCE(SUM(d.amount),0) FROM Donation d WHERE d.project = p) >= p.goalAmount")
+        List<Project> listOfApprovedAndGoalReachedProject(@Param("status") String status, @Param("userId") Long userId);
 }
